move constants in security package
