+++ src/main/java/com/singh/astha/notification/service/security/service/impl/JwtServiceImpl.java
@@ -5,11 +5,11 @@ import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.singh.astha.notification.service.exceptions.ResponseException;
+import com.singh.astha.notification.service.security.utils.Constants;
+import com.singh.astha.notification.service.security.utils.ErrorMessages;
 import com.singh.astha.notification.service.security.dtos.JwtPayload;
 import com.singh.astha.notification.service.security.service.JwtService;
-import com.singh.astha.notification.service.utils.AppProperties;
-import com.singh.astha.notification.service.utils.Constants;
-import com.singh.astha.notification.service.utils.ErrorMessages;
+import com.singh.astha.notification.service.security.utils.AppProperties;
 import lombok.SneakyThrows;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
