+++ src/main/java/com/singh/astha/notification/service/security/utils/Constants.java
@@ -1,0 +1,17 @@
+package com.singh.astha.notification.service.security.utils;
+
+public final class Constants {
+
+    public static final String BEARER = "Bearer ";
+
+    public static final String USER_ID = "userId";
+
+    public static final String ROLES = "roles";
+
+    public static final String CONTENT_TYPE = "Content-Type";
+
+    public static final String APPLICATION_JSON = "application/json";
+
+    public static final String AUTHORIZATION = "Authorization";
+
+}
