+++ src/main/java/com/singh/astha/notification/service/security/utils/ErrorMessages.java
@@ -1,0 +1,12 @@
+package com.singh.astha.notification.service.security.utils;
+
+public final class ErrorMessages {
+
+    public static final String INVALID_TOKEN = "Invalid Token";
+
+    public static final String AUTHORIZATION_HEADER_MUST_NOT_BE_NULL_AND_MUST_START_BE_BEARER = "Authorization Header must not be null and must start be Bearer";
+
+    public static final String AUTHORIZATION_HEADER_MUST_BE_PRESENT = "Authorization header must be present";
+
+    public static final String ACCESS_DENIED = "Access Denied";
+}
