+++ src/main/java/com/singh/astha/notification/service/security/utils/MessageConstants.java
@@ -1,0 +1,6 @@
+package com.singh.astha.notification.service.security.utils;
+
+public final class MessageConstants {
+
+    public static final String UNABLE_TO_SET_USER_PASSWORD_AUTHENTICATION_TOKEN = "Unable to set user name password authentication token";
+}
