+++ src/main/java/com/singh/astha/notification/service/utils/Constants.java
@@ -28,21 +28,11 @@ public final class Constants {
 
     public static final String MESSAGE = "message";
 
-    public static final String NOTIFICATION_SUCCESSFULLY_SENT = "Notification Successfully sent";
-
     public static final String OK = "OK";
 
     public static final String TOKEN_ALREADY_EXISTS = "Token already exists";
 
     public static final String AUTHORIZATION = "Authorization";
-
-    public static final String USER_ID = "userId";
-
-    public static final String ROLES = "roles";
-
-    public static final String CONTENT_TYPE = "Content-Type";
-
-    public static final String APPLICATION_JSON = "application/json";
 
     private Constants() {
     }
