+++ src/main/java/com/singh/astha/notification/service/utils/ErrorMessages.java
@@ -6,16 +6,6 @@ public final class ErrorMessages {
 
     public static final String USER_TOKEN_NOT_EXIST = "User Token does not exist";
 
-    public static final String USER_TOKEN_ALREADY_EXIST = "User Token already exists";
-
-    public static final String INVALID_TOKEN = "Invalid Token";
-
-    public static final String AUTHORIZATION_HEADER_MUST_NOT_BE_NULL_AND_MUST_START_BE_BEARER = "Authorization Header must not be null and must start be Bearer";
-
-    public static final String AUTHORIZATION_HEADER_MUST_BE_PRESENT = "Authorization header must be present";
-
-    public static final String ACCESS_DENIED = "Access Denied";
-
     public static final String TEMPLATE_ALREADY_EXISTS = "Template already exists";
 
     public static final String TEMPLATE_NOT_EXIST = "Template does not exist";
