+++ src/main/java/com/singh/astha/notification/service/utils/MessageConstants.java
@@ -2,8 +2,6 @@ package com.singh.astha.notification.service.utils;
 
 public final class MessageConstants {
 
-    public static final String UNABLE_TO_SET_USER_PASSWORD_AUTHENTICATION_TOKEN = "Unable to set user name password authentication token";
-
     public static final String ENUM_INVALID_CAST_MESSAGE = "must we among %s";
 
     private MessageConstants() {
